requirements.txt
openai==0.27.8
requests==2.20.0
python-inquirer==1.0.3
duckduckgo_search==3.0.2